HW13 Restructure protobuf folder
