+++ hw12_13_14_15_calendar/internal/server/grpc/service.go
@@ -7,7 +7,7 @@ import (
 	"github.com/leksss/hw-test/hw12_13_14_15_calendar/internal/domain/entities"
 	"github.com/leksss/hw-test/hw12_13_14_15_calendar/internal/domain/errors"
 	"github.com/leksss/hw-test/hw12_13_14_15_calendar/internal/domain/interfaces"
-	pb "github.com/leksss/hw-test/hw12_13_14_15_calendar/pb/event"
+	pb "github.com/leksss/hw-test/hw12_13_14_15_calendar/proto/protobuf"
 	"google.golang.org/protobuf/types/known/timestamppb"
 )
